import: added gallery import type
this creates galleries. I creates a post in a given
category together with attachments for all files
in a given folder.
The post contains the [gallery] shortcode so that it displays
a gallery.

This creates symlinks to the files inside the wp_upload_dir, so it may
not work on windows.