- License from AGPL to GPLv2+
I just noticed on FSF website that AGPL is not compatible with GPLv2
(it is with GPLv3 though), and as such is not possible for a Wordpress
plugin if I want it to be on wordpress.org.
So I pass the plugin to GPLv2 and more (so GPLv3 as well).