Don't change locale on frontend AJAX requests
Since all AJAX requests go through wp-admin/admin-ajax.php
they all gets the WP_ADMIN constant and thus are counted as
admin requests by Admin in English.

But sometimes these requets come from the frontend and we
don't want to change the locale on these.

To do this, we check the referer, which the browser sets for all
AJAX requests.