Comments: Smarter color_scheme default
Remove old references to custom CSS

Try to guess a good color_scheme default based on the current theme.

Use the default color scheme blog-side and admin-side if the option isn't set.

Don't store the color_scheme option if it's set to the default so that theme switches and theme option changes will trigger our guesser.