Posterous Importer: Fix a bunch of issues with importing images. We had been pulling in scaled images in a way that wasn't reliable because Posterous seems to sometimes use an infix of "scaled1000" and other times of "scaled.1000" and a simple replacement doesn't work.
Instead of using a regex to pull out a post's images and hoping a simple replacement of the filename works (it doesn't in many cases because of some other gobbledegook in the urls that doesn't always remain the same), use the API's media metadata. This seems to work a lot better.