Jetpack: Publicize
Casting the page_id has an int causes 32-bit systems to always return Facebook IDs as 2147483647. This is the largest int those systems can handle. We don't really need to cast it as an int anyway since we do a digit security check and can just escape for output using esc_attr.