strip trailing slashes from a source url
@boltronics spotted an issue with the SimplePie in WP 3.5 [1]:
if your source url is http://micro.example.com/, the plugin adds a
/rss to it, resulting in http://micro.example.com//rss, which works
with most webservers, but is actually a non valid URI. Since WP 3.5
SimplePie does not ignore this and fucks up (instead of showing a
proper error).

While this will or will not be fixed in SimplePie [2], lets strip the
trailing slash for good, it will be added if needed anyways.

[1] https://github.com/evgeni/wp-statusnet-widget/pull/4
[2] https://github.com/simplepie/simplepie/issues/257