Fix the "Window title" setting not working for anything except custom menus.
Cause: Missed this one when switching to the new replace-and-restore menu modification model. In WP core the code that determines the admin title is separate from the menu rendering algorithm, so it still saw the default menu, not our custom structure. Fixed by making better use of the "admin_title" filter.